+++ StreamFiles/StreamFiles/SFManager.h
@@ -10,13 +10,13 @@
 
 @interface SFManager : NSObject
 
-+(void) readDataFromPath : (NSString*) path withStream : (const void* (^)(uint8_t* buffer, unsigned int length)) stream completion : (void (^)(BOOL isSuccess, NSData *data)) completion;
++(void) readDataFromPath : (NSString*) path withStream : (const void* (^)(uint8_t* buffer, unsigned long length)) stream completion : (void (^)(BOOL isSuccess, NSData *data)) completion;
 
-+(void) writeDataToPath : (NSString*) path withData : (NSData*) data withStream : (const uint8_t* (^)(uint8_t* buffer, unsigned int length)) stream completion : (void (^)(BOOL isSuccess)) completion;
++(void) writeDataToPath : (NSString*) path withData : (NSData*) data withStream : (const uint8_t* (^)(uint8_t* buffer, unsigned long length)) stream completion : (void (^)(BOOL isSuccess)) completion;
 
 +(void) readFromPath : (NSString*) fromPath
          writeToPath : (NSString*) toPath
-          withStream : (const void* (^)(uint8_t* buffer, unsigned int length)) stream
+          withStream : (const void* (^)(uint8_t* buffer, unsigned long length)) stream
           completion : (void (^)(BOOL isSuccess)) completion;
 
 @end
