+++ StreamFiles/StreamFiles/SFOnTheFly.h
@@ -14,7 +14,7 @@
 
 -(void) readFromPath : (NSString*) fromPath
          writeToPath : (NSString*) toPath
-          withStream : (const void* (^)(uint8_t* buffer, unsigned int length)) stream
+          withStream : (const void* (^)(uint8_t* buffer, unsigned long length)) stream
           completion : (void (^)(BOOL isSuccess)) completion;
 
 @end
